+++ python/par_term_emu_core_rust/__init__.py
@@ -84,7 +84,7 @@ except ImportError:
     encode_client_message = None
     decode_client_message = None
 
-__version__ = "0.27.0"
+__version__ = "0.28.0"
 __all__ = [
     "AmbiguousWidth",
     "Attributes",